some condition improvements to remove branches and added capability to draw and run

+++ include/gol.h
@@ -3,5 +3,5 @@
 #include "cell_system.h"
 #include "types.h"
 
-i32 get_total_neighbours(const u16 buffer[TOTAL_CELL_Y][TOTAL_CELL_X], i32 posX, i32 posY);
-void calc_next_buffer(u16 new_buffer[TOTAL_CELL_Y][TOTAL_CELL_X], const u16 old_buffer[TOTAL_CELL_Y][TOTAL_CELL_X]);
+i32 get_total_neighbours(const u8 buffer[TOTAL_CELL_Y][TOTAL_CELL_X], i32 posX, i32 posY);
+void calc_next_buffer(u8 new_buffer[TOTAL_CELL_Y][TOTAL_CELL_X], const u8 old_buffer[TOTAL_CELL_Y][TOTAL_CELL_X]);
